requirements.txt
pandas
numpy
scikit-learn
joblib
python-dateutil
fastapi
uvicorn
requests
flask
faiss-cpu
sentence-transformers
transformers
sqlalchemy
tqdm
python-dotenv